[lint package refactor] Create a file for pylinter
